posix: options: add xsi realtime threads option group
Add a POSIX Option Group XSI_REALTIME_THREADS (with the Kconfig
option CONFIG_XSI_REALTIME_THREADS).

When XSI_REALTIME_THREADS is selected (or when the required POSIX
Options are enabled), define _XOPEN_REALTIME_THREADS to a value
other than -1 (_XOPEN_VERSION was deemed appropriate).
