+++ include/zephyr/posix/posix_features.h
@@ -132,21 +132,7 @@
 #define _POSIX_THREAD_CPUTIME _POSIX_VERSION
 #endif
 
-#ifdef CONFIG_POSIX_THREAD_PRIO_INHERIT
-#define _POSIX_THREAD_PRIO_INHERIT _POSIX_VERSION
-#endif
-
-#ifdef CONFIG_POSIX_THREAD_PRIO_PROTECT
-#define _POSIX_THREAD_PRIO_PROTECT _POSIX_VERSION
-#endif
-
-#ifdef CONFIG_POSIX_THREAD_PRIORITY_SCHEDULING
-#define _POSIX_THREAD_PRIORITY_SCHEDULING _POSIX_VERSION
-#endif
-
 /* #define _POSIX_THREAD_PROCESS_SHARED (-1L) */
-/* #define _POSIX_THREAD_ROBUST_PRIO_INHERIT (-1L) */
-/* #define _POSIX_THREAD_ROBUST_PRIO_PROTECT (-1L) */
 
 /* #define _POSIX_THREAD_SPORADIC_SERVER (-1L) */
 
@@ -214,7 +200,7 @@
 	 defined(CONFIG_POSIX_SHARED_MEMORY_OBJECTS) && defined(CONFIG_POSIX_SYNCHRONIZED_IO))
 #define _XOPEN_REALTIME _XOPEN_VERSION
 #endif
-/* #define _XOPEN_REALTIME_THREADS (-1L) */
+
 /* #define _XOPEN_SHM (-1L) */
 
 #ifdef CONFIG_XOPEN_STREAMS
