+++ lib/posix/Kconfig.toolchain
@@ -213,6 +213,9 @@ config TC_PROVIDES_XSI_MULTI_PROCESS
 config TC_PROVIDES_XSI_REALTIME
 	bool
 
+config TC_PROVIDES_XSI_REALTIME_THREADS
+	bool
+
 config TC_PROVIDES_XSI_SIGNALS
 	bool
