+++ lib/posix/options/Kconfig
@@ -30,6 +30,7 @@ menu "X/Open system interfaces"
 
 # zephyr-keep-sorted-start
 rsource "Kconfig.xsi_realtime"
+rsource "Kconfig.xsi_realtime_threads"
 rsource "Kconfig.xsi_single_process"
 rsource "Kconfig.xsi_streams"
 rsource "Kconfig.xsi_system_logging"
