+++ tests/posix/common/prj.conf
@@ -3,7 +3,6 @@ CONFIG_POSIX_THREAD_THREADS_MAX=6
 CONFIG_ZTEST=y
 CONFIG_POSIX_SEM_VALUE_MAX=32767
 CONFIG_POSIX_MESSAGE_PASSING=y
-CONFIG_POSIX_PRIORITY_SCHEDULING=y
 CONFIG_HEAP_MEM_POOL_SIZE=4096
 CONFIG_THREAD_NAME=y
 CONFIG_DYNAMIC_THREAD=y
@@ -26,3 +25,5 @@ CONFIG_XSI_SYSTEM_LOGGING=y
 
 # for sched_get_priority_min(), sched_get_priority_max()
 CONFIG_POSIX_PRIORITY_SCHEDULING=y
+
+CONFIG_XSI_REALTIME_THREADS=y
